extracted component AddNewProduct
