+++ frontend/src/pages/Admin.jsx
@@ -1,34 +1,17 @@
 import React, { useState } from "react";
 import { POST_REQUEST } from "@/utils/helpers/request.helper";
 import * as shad from "@/components/ui/shadBarrel";
+import AddNewProduct from "@/components/admin/AddNewProduct";
 
 export default function Admin() {
-  const [newProduct, setNewProduct] = useState({});
+  const [showAddProduct, setShowAddProduct] = useState(false);
 
-  function handleProductNameChange(e) {
-    setNewProduct({ ...newProduct, name: e.target.value });
-  }
-
-  function handleProductPriceChange(e) {
-    setNewProduct({ ...newProduct, price: e.target.value });
-  }
-
-  function handleProductStockChange(e) {
-    setNewProduct({ ...newProduct, stock: e.target.value });
-  }
-
-  async function handleSubmit(e) {
-    e.preventDefault();
+  async function handleSubmit(newProduct) {
     console.log("test");
 
-    const updateObject = newProduct;
-
-    console.log(updateObject);
+    console.log(newProduct);
     try {
-      const response = await POST_REQUEST(
-        "/api/products/create/",
-        updateObject
-      );
+      const response = await POST_REQUEST("/api/products/create/", newProduct);
       if (response.status === 201) {
         console.log("Good job dude");
       } else {
@@ -39,39 +22,12 @@ export default function Admin() {
     }
   }
 
+  //Lägg in en knapp för att hämta alla produkter
+  // klicka på produkt och få upp en modal med info om produkten
+
   return (
     <>
-      <form
-        onSubmit={async (e) => {
-          await handleSubmit(e);
-        }}
-      >
-        <label htmlFor="productName"> Input product Name</label>
-        <input
-          id="productName"
-          type="text"
-          value={newProduct.name}
-          onChange={handleProductNameChange}
-        />
-        <br />
-        <label htmlFor="productPrice">Product price</label>
-        <input
-          id="productPrice"
-          type="text"
-          value={newProduct.price}
-          onChange={handleProductPriceChange}
-        />{" "}
-        <br />
-        <label htmlFor="productStock">Product stock: </label>
-        <input
-          id="productStock"
-          type="text"
-          value={newProduct.stock}
-          onChange={handleProductStockChange}
-        />{" "}
-        <br />
-        <shad.Button type={"submit"}>Post user</shad.Button>
-      </form>
+      <AddNewProduct onSubmit={handleSubmit} />
     </>
   );
 }
